requirements.txt
# Lightweight ModelOps - Core Dependencies
# For quick setup without Poetry

# Web Framework
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML Frameworks (Lightweight)
torch==2.6.0
transformers==4.36.0
peft==0.7.0
bitsandbytes==0.41.0
trl==0.7.0

# Inference (Laptop-friendly)
llama-cpp-python==0.2.0

# Quantization (CPU-supported)
autoawq==0.1.6

# Orchestration (Lightweight)
prefect==2.14.0

# Storage (Embedded)
duckdb==0.9.0
lancedb==0.4.0

# Vector & Embeddings
sentence-transformers==2.2.0
fastembed==0.1.0

# Experiment Tracking (Local)
mlflow==2.9.0

# Data Processing
pandas==2.1.0
pyarrow==14.0.0
datasets==2.15.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.0
aiofiles==23.2.0
python-multipart==0.0.6
prometheus-client==0.19.0

# CLI
click==8.1.0
rich==13.7.0
typer==0.9.0

# Security & Authentication
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Monitoring & Drift Detection
scipy==1.11.4
numpy==1.26.2
scikit-learn==1.3.2

# Alerting
slack-sdk==3.26.1
sendgrid==6.11.0

# WebSocket for real-time streaming
websockets==12.0

# Dev Dependencies
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.11.0
ruff==0.1.6
requests==2.31.0
